Refactor figure model to include canvas width and height from canvas image metadata
